refactor: 결제 요청 승인 로직 변경

+++ server/module-core/src/main/java/com/team33/modulecore/item/domain/repository/ItemCommandRepository.java
@@ -37,6 +37,6 @@ public interface ItemCommandRepository extends Repository<Item, Long> {
 	 */
 	@Modifying(clearAutomatically = true, flushAutomatically = true)
 	@Query("UPDATE Item i SET i.statistics.sales = i.statistics.sales + 1 WHERE i.id = :itemId")
-	Long incrementSales(@Param("itemId") Long itemId);
+	int incrementSales(@Param("itemId") Long itemId);
 }
